+++ apps/Conduit-UI/src/redux/slices/authenticationSlice.ts
@@ -3,12 +3,9 @@ import { AuthUser, IAuthenticationConfig } from '../../models/authentication/Aut
 import {
   blockUnblockUsers,
   blockUser,
-  changePassword,
   createNewUsers,
-  deleteAdmin,
   deleteUsers,
   editUser,
-  getAdmins,
   getAuthenticationConfig,
   getAuthUsersDataReq,
   putAuthenticationConfig,
@@ -25,10 +22,6 @@ interface IAuthenticationSlice {
       users: AuthUser[];
       count: number;
     };
-    authAdmins: {
-      admins: AuthUser[];
-      count: number;
-    };
     config: IAuthenticationConfig;
   };
 }
@@ -37,10 +30,6 @@ const initialState: IAuthenticationSlice = {
   data: {
     authUsers: {
       users: [],
-      count: 0,
-    },
-    authAdmins: {
-      admins: [],
       count: 0,
     },
     config: {
@@ -278,55 +267,6 @@ export const asyncUpdateAuthenticationConfig = createAsyncThunk(
       thunkAPI.dispatch(setAppLoading(false));
       thunkAPI.dispatch(enqueueSuccessNotification(`Authentication config successfully updated`));
       return config;
-    } catch (error) {
-      thunkAPI.dispatch(setAppLoading(false));
-      thunkAPI.dispatch(enqueueErrorNotification(`${getErrorData(error)}`));
-      throw error;
-    }
-  }
-);
-
-export const asyncGetAdmins = createAsyncThunk(
-  'authentication/getAdmins',
-  async (params: Pagination, thunkAPI) => {
-    thunkAPI.dispatch(setAppLoading(true));
-    try {
-      const { data } = await getAdmins(params);
-
-      console.log(data);
-      thunkAPI.dispatch(setAppLoading(false));
-    } catch (error) {
-      thunkAPI.dispatch(setAppLoading(false));
-      thunkAPI.dispatch(enqueueErrorNotification(`${getErrorData(error)}`));
-      throw error;
-    }
-  }
-);
-
-export const asyncDeleteAdmin = createAsyncThunk(
-  'authentication/deleteAdmin',
-  async (params: { id: string; getAdmins: any }, thunkAPI) => {
-    thunkAPI.dispatch(setAppLoading(true));
-    try {
-      const { data } = await deleteAdmin(params.id);
-      params.getAdmins();
-      thunkAPI.dispatch(setAppLoading(false));
-    } catch (error) {
-      thunkAPI.dispatch(setAppLoading(false));
-      thunkAPI.dispatch(enqueueErrorNotification(`${getErrorData(error)}`));
-      throw error;
-    }
-  }
-);
-
-export const asyncChangePassword = createAsyncThunk(
-  'authentication/changePassword',
-  async (params: { newPassword: string; oldPassword: string }, thunkAPI) => {
-    thunkAPI.dispatch(setAppLoading(true));
-    try {
-      const { data } = await changePassword(params.newPassword, params.oldPassword);
-
-      thunkAPI.dispatch(setAppLoading(false));
     } catch (error) {
       thunkAPI.dispatch(setAppLoading(false));
       thunkAPI.dispatch(enqueueErrorNotification(`${getErrorData(error)}`));
